requirements.txt
numpy

uvicorn
pyjwt
fastapi

# Database
sqlmodel
pydantic-settings
pymysql

# OAuth
pyjwt
pwdlib